Refactor record module packages
