+++ app/src/main/kotlin/com/gdavidpb/tuindice/di/AppModule.kt
@@ -60,10 +60,10 @@ import com.gdavidpb.tuindice.presentation.action.main.StartUpActionProcessor
 import com.gdavidpb.tuindice.presentation.action.main.UpdateStateActionProcessor
 import com.gdavidpb.tuindice.presentation.viewmodel.BrowserViewModel
 import com.gdavidpb.tuindice.presentation.viewmodel.MainViewModel
-import com.gdavidpb.tuindice.record.data.api.parser.QuarterRemoveParser
-import com.gdavidpb.tuindice.record.data.api.parser.SubjectUpdateParser
-import com.gdavidpb.tuindice.record.data.room.resolution.QuarterResolutionHandler
-import com.gdavidpb.tuindice.record.data.room.resolution.SubjectResolutionHandler
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.parser.QuarterRemoveParser
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.parser.SubjectUpdateParser
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.database.QuarterResolutionHandler
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.database.SubjectResolutionHandler
 import com.gdavidpb.tuindice.transactions.data.api.transaction.TransactionInterceptor
 import com.gdavidpb.tuindice.transactions.data.api.transaction.TransactionParser
 import com.gdavidpb.tuindice.transactions.data.room.resolution.ResolutionApplier
