+++ evaluations/src/main/kotlin/com/gdavidpb/tuindice/evaluations/data/repository/evaluation/source/api/mapper/Evaluation.kt
@@ -9,7 +9,7 @@ import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationAdd
 import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationUpdate
 import com.gdavidpb.tuindice.evaluations.utils.extension.computeEvaluationState
 import com.gdavidpb.tuindice.persistence.data.room.entity.EvaluationEntity
-import com.gdavidpb.tuindice.record.data.api.mapper.toSubject
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.mapper.toSubject
 import java.util.Date
 
 fun EvaluationResponse.toEvaluation() = Evaluation(
