+++ evaluations/src/main/kotlin/com/gdavidpb/tuindice/evaluations/data/repository/evaluation/source/api/response/EvaluationResponse.kt
@@ -1,6 +1,6 @@
 package com.gdavidpb.tuindice.evaluations.data.repository.evaluation.source.api.response
 
-import com.gdavidpb.tuindice.record.data.api.response.SubjectResponse
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.response.SubjectResponse
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
