+++ record/src/main/kotlin/com/gdavidpb/tuindice/record/data/repository/quarter/source/RecordApiDataSource.kt
@@ -1,16 +1,17 @@
-package com.gdavidpb.tuindice.record.data.api
+package com.gdavidpb.tuindice.record.data.repository.quarter.source
 
 import com.gdavidpb.tuindice.base.domain.model.quarter.Quarter
 import com.gdavidpb.tuindice.base.domain.model.subject.Subject
 import com.gdavidpb.tuindice.base.utils.extension.getOrThrow
-import com.gdavidpb.tuindice.record.data.api.mapper.toQuarter
-import com.gdavidpb.tuindice.record.data.api.mapper.toSubject
-import com.gdavidpb.tuindice.record.data.api.mapper.toUpdateSubjectRequest
-import com.gdavidpb.tuindice.record.data.quarter.source.RemoteDataSource
+import com.gdavidpb.tuindice.record.data.repository.quarter.RecordApi
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.mapper.toQuarter
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.mapper.toSubject
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.mapper.toUpdateSubjectRequest
+import com.gdavidpb.tuindice.record.data.repository.quarter.RemoteDataSource
 import com.gdavidpb.tuindice.record.domain.model.QuarterRemove
 import com.gdavidpb.tuindice.record.domain.model.SubjectUpdate
 
-class ApiDataSource(
+class RecordApiDataSource(
 	private val recordApi: RecordApi
 ) : RemoteDataSource {
 	override suspend fun getQuarters(): List<Quarter> {
