+++ record/src/main/kotlin/com/gdavidpb/tuindice/record/data/repository/quarter/source/api/parser/QuarterRemoveParser.kt
@@ -1,4 +1,4 @@
-package com.gdavidpb.tuindice.record.data.api.parser
+package com.gdavidpb.tuindice.record.data.repository.quarter.source.api.parser
 
 import com.gdavidpb.tuindice.base.domain.model.transaction.Transaction
 import com.gdavidpb.tuindice.base.domain.model.transaction.TransactionAction
