+++ record/src/main/kotlin/com/gdavidpb/tuindice/record/data/repository/quarter/source/api/response/QuarterResponse.kt
@@ -1,4 +1,4 @@
-package com.gdavidpb.tuindice.record.data.api.response
+package com.gdavidpb.tuindice.record.data.repository.quarter.source.api.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
