+++ record/src/main/kotlin/com/gdavidpb/tuindice/record/data/repository/quarter/source/database/SubjectResolutionHandler.kt
@@ -1,46 +1,41 @@
-package com.gdavidpb.tuindice.record.data.room.resolution
+package com.gdavidpb.tuindice.record.data.repository.quarter.source.database
 
 import com.gdavidpb.tuindice.base.domain.model.resolution.Resolution
 import com.gdavidpb.tuindice.base.domain.model.resolution.ResolutionAction
 import com.gdavidpb.tuindice.base.domain.model.resolution.ResolutionType
 import com.gdavidpb.tuindice.persistence.data.room.TuIndiceDatabase
-import com.gdavidpb.tuindice.record.data.api.mapper.toQuarterAndSubjectsEntities
-import com.gdavidpb.tuindice.record.data.api.response.QuarterResponse
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.mapper.toSubjectEntity
+import com.gdavidpb.tuindice.record.data.repository.quarter.source.api.response.SubjectResponse
 import com.gdavidpb.tuindice.transactions.data.room.resolution.ResolutionHandler
 import kotlinx.serialization.json.Json
 
-class QuarterResolutionHandler(
+class SubjectResolutionHandler(
 	private val room: TuIndiceDatabase
 ) : ResolutionHandler {
 	override suspend fun match(resolution: Resolution): Boolean {
-		return (resolution.type == ResolutionType.QUARTER)
+		return (resolution.type == ResolutionType.SUBJECT)
 	}
 
 	override suspend fun apply(resolution: Resolution) {
-		val (quarterEntity, subjectEntities) = Json.decodeFromString<QuarterResponse>(string = resolution.data)
-			.toQuarterAndSubjectsEntities(uid = resolution.uid)
+		val subjectEntity = Json.decodeFromString<SubjectResponse>(string = resolution.data)
+			.toSubjectEntity(uid = resolution.uid)
 
 		if (resolution.localReference != resolution.remoteReference)
-			room.quarters.updateId(
+			room.subjects.updateId(
 				fromId = resolution.localReference,
 				toId = resolution.remoteReference
 			)
 
 		when (resolution.action) {
-			ResolutionAction.ADD, ResolutionAction.UPDATE -> {
-				room.quarters.upsertEntities(
-					entities = listOf(quarterEntity)
-				)
-
+			ResolutionAction.ADD, ResolutionAction.UPDATE ->
 				room.subjects.upsertEntities(
-					entities = subjectEntities
+					entities = listOf(subjectEntity)
 				)
-			}
 
 			ResolutionAction.DELETE ->
-				room.quarters.deleteQuarter(
-					uid = quarterEntity.accountId,
-					qid = quarterEntity.id
+				room.subjects.deleteSubject(
+					uid = subjectEntity.accountId,
+					sid = subjectEntity.id
 				)
 		}
 	}
